+++ CHANGELOG.md
@@ -2,6 +2,8 @@ v0.2.0 (Develop)
 ---
 * Added `:filename` regex.
 * removed unnecesary `print()` statement.
+* Added *pep8* recommendations.
+* Added `UUID` pattern.
 
 v0.1.3
 ---